fix: move captureWarningMessageFromError
